There is a multithreading issue regarding the ModelRemoteGenerationWorker which is instantiated in ModelGen...Service. Temporary fix applied, no exceptions thrown.
