+++ subprojects/language-web/src/main/java/tools/refinery/language/web/generator/ModelGenerationService.java
@@ -22,7 +22,7 @@ public class ModelGenerationService {
 	public static final String SERVICE_NAME = "modelGeneration";
 	public static final String MODEL_GENERATION_EXECUTOR = "modelGeneration";
 	public static final String MODEL_GENERATION_TIMEOUT_EXECUTOR = "modelGenerationTimeout";
-	public static final boolean USE_GENERATOR_SERVER = true;
+	public static  boolean USE_GENERATOR_SERVER = true;
 
 	@Inject
 	private OperationCanceledManager operationCanceledManager;
@@ -37,6 +37,9 @@ public class ModelGenerationService {
 
 	public ModelGenerationService() {
 		timeoutSec = SemanticsService.getTimeout("REFINERY_MODEL_GENERATION_TIMEOUT_SEC").orElse(600L);
+		String useGenServer = System.getenv("USE_GENERATOR_SERVER");
+		if (useGenServer != null)
+			USE_GENERATOR_SERVER = Boolean.getBoolean(useGenServer);
 	}
 
 	private ModelGenerationStartedResult generateModelLocal(PushWebDocumentAccess document, int randomSeed){
@@ -66,8 +69,8 @@ public class ModelGenerationService {
 				var worker = remoteWorkerProvider.get();
 				worker.setState(pushState, randomSeed, timeoutSec);
 				var manager = pushState.getModelGenerationManager();
-				worker.start();
 				boolean canceled = manager.setActiveModelGenerationWorker(worker, cancelIndicator);
+				worker.start();
 				if (canceled) {
 					worker.cancel();
 					operationCanceledManager.throwOperationCanceledException();
